requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==10.0.0
django-crispy-forms==1.14.0
googlemaps==4.10.0
whitenoise==6.4.0
django-environ==0.9.0
pyodbc==5.2.0
mssql-django==1.5